drunkeness reduces automatically, re-arranged bottle position, back button in item shop, room name in itemshop

+++ Source/ItemManager.h
@@ -23,7 +23,7 @@ enum ItemType {
     POWERUP = 3,
 };
 
-+ (int) checkBeerBottles : (CCSprite*) dave :(CCSprite*) huey :(int*) daveDrunkLevel :(int*) hueyDrunkLevel :(__strong CCNode*[]) beerNodes;
++ (int) checkBeerBottles : (CCSprite*) dave :(CCSprite*) huey :(float*) daveDrunkLevel :(float*) hueyDrunkLevel :(__strong CCNode*[]) beerNodes;
 + (CCNode*)dropItem;;
 + (CGPoint)itemDisplay;
 + (int) useItem: (__strong CCNode*[]) itemBoxes : (int) index : (int) itemsHeld;
